pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "publishes my resume"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jinja2>=3.1.6",
    "markdown-it-py>=4.0.0",
    "weasyprint>=68.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
